+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 * Return original unwrapped handler from `script.get_event_handler`
 * Fixed various missing stepIgnores
 * stepIgnore and stepIgnoreAll now return the function/table passed in, for convenience
-* Objects printed with `__DebugAdapter.print(obj)` will be expandable in debug console until the next break-and-continue or until the current lua stack ends
+* Objects printed with `__DebugAdapter.print(obj)` will be expandable in debug console as long as the source Lua State is still active
 * Pause button able to break into long running code
 * General code cleanup
 * Optional arguments `upStack` and `category` on `__DebugAdapter.print()`